ui location fix

+++ profile.php
@@ -311,6 +311,7 @@ body {
 			$content = $row_posts['post_content'];
 			$upload_image = $row_posts['upload_image'];
 			$post_date = $row_posts['post_date'];
+            $location = $row_posts['location'];
 
 			$user = "SELECT * FROM `users` WHERE user_id='$user_id' AND posts='yes'";
 			$run_user = mysqli_query($con,$user);
@@ -321,7 +322,7 @@ body {
 
 			//display post
 
-			if($content == "No" && strlen($upload_image) >= 1)
+			if($content == "No" && strlen($upload_image) >= 1 && strlen($location) >= 1)
 			{
 				echo "
 				<div id='own_posts'>
@@ -329,11 +330,10 @@ body {
 					<div class='col-sm-2'>
 						<p><img src='users/$user_image' class='rounded-circle' width='100px' height='100px'></p>
 					</div>
-					<div class='col-sm-6'>
+					<div class='col'>
 						<h5><a style='text-decoration: none;cursor: pointer;color: #3897f0;' href='user_profile.php?u_id=$user_id'>$user_name</a></h5>
 						<p><small style='color:black;'>Updated a post on <strong>$post_date <i class='fas fa-globe-asia'></i></strong></small><p>
-					</div>
-					<div class='col-sm-4'>
+                        <h6><small style='color:#9e9e9e;font-size: 69%;'><strong><i class='fas fa-location-arrow'></i> $location </strong></small></h6>
 					</div>
 				</div>
 				<div class='row'>
@@ -347,7 +347,31 @@ body {
 			</div><br><br>
 				";
 			}
-			else if(strlen($content) >= 1 && strlen($upload_image) >= 1){
+            else if($content == "No" && strlen($upload_image) >= 1 && strlen($location) == 0)
+            {
+                echo "
+                <div id='own_posts'>
+                <div class='row'>
+                    <div class='col-sm-2'>
+                        <p><img src='users/$user_image' class='rounded-circle' width='100px' height='100px'></p>
+                    </div>
+                    <div class='col'>
+                        <h5><a style='text-decoration: none;cursor: pointer;color: #3897f0;' href='user_profile.php?u_id=$user_id'>$user_name</a></h5>
+                        <p><small style='color:black;'>Updated a post on <strong>$post_date <i class='fas fa-globe-asia'></i></strong></small><p>
+                    </div>
+                </div>
+                <div class='row'>
+                    <div class='col-sm-12'>
+                        <img class='img-fluid' id='posts-img' src='imagepost/$upload_image' style='height:200px;'>
+                    </div>
+                </div><br>
+                <a href='single.php?post_id=$post_id' style='float:right;'><button class='btn btn-success btn-sm'>View</button></a>
+                <a href='edit_post.php?post_id=$post_id' style='float:right;'><button  class='btn btn-info btn-sm'>Edit</button></a>
+                <a href='functions/delete_post.php?post_id=$post_id' class='btn btn-danger btn-sm' style='float:right;'>Delete Post</a><br>
+            </div><br><br>
+                ";
+            }
+			else if(strlen($content) >= 1 && strlen($upload_image) >= 1 && strlen($location) >= 1){
 			echo "
 			<div id='own_posts'>
 				<div class='row'>
@@ -356,12 +380,10 @@ body {
 					<div class='col-sm-2'>
 						<p><img src='users/$user_image' class='rounded-circle' width='100px' height='100px'></p>
 					</div>
-					<div class='col-sm-6'>
+					<div class='col'>
 						<h5><a style='text-decoration: none;cursor: pointer;color: #3897f0;' href='user_profile.php?u_id=$user_id'>$user_name</a></h5>
 						<p><small style='color:black;'>Updated a post on <strong>$post_date <i class='fas fa-globe-asia'></i></strong></small></p>
-					</div>
-					<div class='col-sm-4'>
-						
+                            <h6><small style='color:#9e9e9e;font-size: 69%;'><strong><i class='fas fa-location-arrow'></i> $location </strong></small></h6>
 					</div>
 				</div>
 				<div class='row'>
@@ -376,20 +398,19 @@ body {
 			</div><br><br>
 			";
 			}
-			else
+			else if(strlen($content) >= 1)
 			{
+                if(strlen($location) >= 1 && strlen($content) >= 1) {
 			echo "
 			<div id='own_posts'>
 				<div class='row'>
 					<div class='col-sm-2'>
 						<p><img src='users/$user_image' class='rounded-circle' width='100px' height='100px'></p>
 					</div>
-					<div class='col-sm-6'>
+					<div class='col'>
 						<h5><a style='text-decoration: none;cursor: pointer;color: #3897f0;' href='user_profile.php?u_id=$user_id'>$user_name</a></h5>
 						<p><small style='color:black;'>Updated a post on <strong>$post_date <i class='fas fa-globe-asia'></i></strong></small></p>
-					</div>
-					<div class='col-sm-4'>
-						
+                        <h6><small style='color:#9e9e9e;font-size: 69%;'><strong><i class='fas fa-location-arrow'></i> $location </strong></small></h6>
 					</div>
 				</div>
 				<div class='row'>
@@ -399,9 +420,13 @@ body {
 						<label>$content</label>
 					</div>
 					<div class='col-sm-4'>
-						
 					</div>
+                    </div>
+                    <a href='single.php?post_id=$post_id' style='float:right;'><button class='btn btn-success btn-sm'>View</button></a>
+                <a href='edit_post.php?post_id=$post_id' style='float:right;'><button  class='btn btn-info btn-sm'>Edit</button></a>
+                <a href='functions/delete_post.php?post_id=$post_id' class='btn btn-danger btn-sm' style='float:right;'>Delete Post</a><br>
 				</div><br><br>
+
 			";
 			global $con;
 	
@@ -429,14 +454,70 @@ body {
 			else
 			{
 			echo"
-				<a href='single.php?post_id=$post_id' style='float:right;'><button class='btn btn-success btn-sm'>View</button></a>
-				<a href='edit_post.php?post_id=$post_id' style='float:right;'><button  class='btn btn-info btn-sm'>Edit</button></a>
-				<a href='functions/delete_post.php?post_id=$post_id' class='btn btn-danger btn-sm' style='float:right;'>Delete Post</a><br>
-			</div><br><br>
+				
+			
 			";
 			}
+          }
+           else if(strlen($location) == 0) {
+            echo"
+                <div class='row'>
+                    <div id='own_posts' class='col-md-12'>
+                        <div class='row'>
+                            <div class='col-sm-2'>
+                            <p><img src='users/$user_image' class='rounded-circle' width='100px' height='100px'></p>
+                            </div>
+                            <div class='col'>
+                                <h3><a style='text-decoration:none; cursor:pointer;color #3897f0;' href='user_profile.php?u_id=$user_id'>$user_name</a></h3>
+                                <h5><small style='color:black;'>Updated a post on <strong>$post_date </strong><i class='fas fa-globe-asia'></i></small></h5>
+                            </div>
+                        </div>
+                        <div class='row'>
+                            <div class='col-sm-12'>
+                                <h3 style='color:black;'>$content</h3>
+                            </div>
+                        </div><br>
+                        <a href='single.php?post_id=$post_id' style='float:right;'><button class='btn btn-success btn-sm'>View</button></a>
+                <a href='edit_post.php?post_id=$post_id' style='float:right;'><button  class='btn btn-info btn-sm'>Edit</button></a>
+                <a href='functions/delete_post.php?post_id=$post_id' class='btn btn-danger btn-sm' style='float:right;'>Delete Post</a><br>
+                    </div>
+                    <div class='col-sm-3'>
+                    </div>
+
+                </div><hr>
+                ";
+           }
 		}
-	}
+	
+    else if(strlen($content) == 0 && strlen($upload_image) == 0 && strlen($location) >= 1) {
+        echo "
+            <div id='own_posts'>
+                <div class='row'>
+                    <div class='col-sm-2'>
+                        <p><img src='users/$user_image' class='rounded-circle' width='100px' height='100px'></p>
+                    </div>
+                    <div class='col'>
+                        <h5><a style='text-decoration: none;cursor: pointer;color: #3897f0;' href='user_profile.php?u_id=$user_id'>$user_name</a></h5>
+                        <p><small style='color:black;'>Check-in on <strong>$post_date <i class='fas fa-globe-asia'></i></strong></small></p>
+                        <h5><small style='color:#9e9e9e;font-size: 69%;'><strong><i class='fas fa-location-arrow'></i> $location </strong></small></h5>
+                    </div>
+                </div>
+                <div class='row'>
+                    <div class='col-sm-2'>
+                    </div>
+                    <div class='col-sm-6'>
+                        <label></label>
+                    </div>
+                    <div class='col-sm-4'>
+                        
+                    </div>
+                </div>
+                <a href='single.php?post_id=$post_id' style='float:right;'><button class='btn btn-success btn-sm'>View</button></a>
+                <a href='functions/delete_post.php?post_id=$post_id' class='btn btn-danger btn-sm' style='float:right;'>Delete Post</a><br>
+                </div><br><br>
+            ";
+    }
+}
 		?>      
                             </div>
                         </div>
